requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
DateTime==5.5
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.1.2
pandas==2.2.3
psycopg2==2.9.10
pygame==2.6.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
zope.interface==7.1.1
